refactor: Convert utils/collection to Typescript (#23992)
